+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
-- `add_variables.py`: support for 3/4 time without denominator
+- `add_variables.py`:
+  - support for 3/4 time without denominator
+  - support for a different displayed time signature
 - [Pixi](https://pixi.prefix.dev) as environment manager